rename: change jsx to tsx
