+++ src/pages/additem/AddItemPage.tsx
@@ -1,5 +1,5 @@
 function AddItemPage() {
-    return <h2>임시 페이지</h2>;
+  return <h2>임시 페이지</h2>;
 }
 
 export default AddItemPage;
